add support for B3 in FieldWithHelpBubble and move into style/crispy.py
