+++ corehq/apps/style/templates/style/crispy/bootstrap/field_with_help_bubble.html
@@ -1,0 +1,48 @@
+{% load crispy_forms_field %}
+
+{% if field.is_hidden %}
+    {{ field }}
+{% else %}
+    <{% if tag %}{{ tag }}{% else %}div{% endif %}
+        id="div_{{ field.auto_id }}"
+        class="control-group{% if wrapper_class %} {{ wrapper_class }}{% endif %}{% if form_show_errors%}{% if field.errors %} error{% endif %}{% endif %}{% if field.css_classes %} {{ field.css_classes }}{% endif %}"
+    >
+        {% if field.label and not field|is_checkbox and form_show_labels %}
+            <div class="control-label {% if field.field.required %}requiredField{% endif %}">
+                <label for="{{ field.id_for_label }}" class="inner-control-label">
+                    {{ field.label|safe }}{% if field.field.required %}<span class="asteriskField">*</span>{% endif %}
+                </label>
+                {% include 'style/crispy/bootstrap/help_bubble.html' %}
+            </div>
+        {% endif %}
+
+        {% if field|is_checkboxselectmultiple %}
+            {% include 'bootstrap/layout/checkboxselectmultiple.html' %}
+        {% endif %}
+
+        {% if field|is_checkbox and help_bubble_text %}
+            <label class="control-label">
+                {% include 'style/crispy/bootstrap/help_bubble.html' %}
+            </label>
+        {% endif %}
+
+        {% if field|is_radioselect %}
+            {% include 'bootstrap/layout/radioselect.html' %}
+        {% endif %}
+
+        {% if not field|is_checkboxselectmultiple and not field|is_radioselect %}
+            <div class="controls">
+                {% if field|is_checkbox and form_show_labels %}
+                    <label for="{{ field.id_for_label }}" class="checkbox {% if field.field.required %}requiredField{% endif %}">
+                        {% crispy_field field %}
+                        {{ field.label|safe }}
+                        {% include 'bootstrap/layout/help_text_and_errors.html' %}
+                    </label>
+                {% else %}
+                    {% crispy_field field %}
+                    {% include 'bootstrap/layout/help_text_and_errors.html' %}
+                {% endif %}
+            </div>
+        {% endif %}
+    </{% if tag %}{{ tag }}{% else %}div{% endif %}>
+{% endif %}
