+++ corehq/apps/style/templates/style/crispy/bootstrap3/help_bubble.html
@@ -1,0 +1,14 @@
+{% if help_bubble_text %}
+    <span class="hq-help-template"
+        data-title="{{ field.label }}"
+        data-content="{{ help_bubble_text }}"
+        data-bootstrap3="true"
+    ></span>
+    <script>
+        $(function () {
+            $('.hq-help-template').each(function () {
+                COMMCAREHQ.transformHelpTemplate($(this), true);
+            });
+        });
+    </script>
+{% endif %}
